+++ packages/forma-36-tokens/src/tokens/colors/colors-red-deprecated.js
@@ -1,0 +1,11 @@
+const scaleRed = require('./scale-red');
+
+const colorsRed = {
+  'color-red-dark': scaleRed['red-700'],
+  'color-red-base': scaleRed['red-600'],
+  'color-red-mid': scaleRed['red-500'],
+  'color-red-light': scaleRed['red-300'],
+  'color-red-lightest': scaleRed['red-100'],
+};
+
+module.exports = colorsRed;
