+++ packages/forma-36-tokens/src/tokens/colors/scale-red.js
@@ -1,0 +1,13 @@
+const scaleRed = {
+  'red-100': '#FFF2F2',
+  'red-200': '#FFE0E0',
+  'red-300': '#FFB1B2',
+  'red-400': '#FF707D',
+  'red-500': '#DA294A',
+  'red-600': '#BD002A',
+  'red-700': '#990017',
+  'red-800': '#7F0010',
+  'red-900': '#72000E',
+};
+
+module.exports = scaleRed;
